Recent Activity Section
Setup HTML framework for Recent Activity section using the user
information from the Random User Generator API. Mirrored the styling
and structure of the New Members section. Also, updated javascript to
populate the HTML by modifying the Member HTML function to handle both
sections. Further changes are needed to the HTML and spacing is an
issue.