Message User Section
Added Message User section and simulated a message being send when the
user clicks on the send button by providing a confirmation notice. An
error notice will be added later if the user tries to send a message
without filling in the both fields. Also, incorporated the clearfix
helper class to prevent the parent of the floated New Members and
Recent Activity sections from collapsing.