Navigation Bar
Made the width of the navigation bar equal to the height of the header.
Added color line to the left of the “selected” icon to show which page
is active. Added title attributes to the nav icons to identify them
more clearly. Removed mouseover event that added tooltips to paragraphs
in the Recent Activity section that was truncated because it failed to
fire consistently.