Fixes
Thanks to @sofian for finding the reason why my project wouldn’t scale
down to 320px on Firefox (flex: 1 on main container). Also, made a
change to the toggle switch label width which was slightly too long for
the charts breakpoint. Finally, adjusted the font weight down on the
message sent and message error notification.